+++ test/nodes/modbus-queue-info-test.js
@@ -88,10 +88,10 @@ describe('Queue Info node Testing', function () {
           topic: '',
           payload: '',
           payloadType: 'date',
-          repeat: '0.2',
+          repeat: 1,
           crontab: '',
           once: true,
-          onceDelay: 0.1,
+          onceDelay: 1,
           wires: [
             [
               'ef5dad20.e97af'
@@ -137,7 +137,7 @@ describe('Queue Info node Testing', function () {
       })
     })
 
-    it('simple flow with inject should be loaded', function (done) {
+    it('simple flow with old reset inject should be loaded', function (done) {
       helper.load([injectNode, clientNode, serverNode, nodeUnderTest], [{
         id: '445454e4.968564',
         type: 'modbus-server',
@@ -183,10 +183,10 @@ describe('Queue Info node Testing', function () {
         topic: '',
         payload: '',
         payloadType: 'date',
-        repeat: '0.1',
+        repeat: 2,
         crontab: '',
-        once: true,
-        onceDelay: 0.1,
+        once: false,
+        onceDelay: 1,
         wires: [
           [
             '5fffb0bc.0b8a5'
@@ -224,6 +224,98 @@ describe('Queue Info node Testing', function () {
         })
         const queueNode = helper.getNode('5fffb0bc.0b8a5')
         queueNode.receive({ payload: '', resetQueue: true })
+      }, function () {
+        helper.log('function callback')
+      })
+    })
+
+    it('simple flow with new reset inject should be loaded', function (done) {
+      helper.load([injectNode, clientNode, serverNode, nodeUnderTest], [{
+        id: '445454e4.968564',
+        type: 'modbus-server',
+        name: '',
+        logEnabled: true,
+        hostname: '127.0.0.1',
+        serverPort: '7502',
+        responseDelay: 100,
+        delayUnit: 'ms',
+        coilsBufferSize: 10000,
+        holdingBufferSize: 10000,
+        inputBufferSize: 10000,
+        discreteBufferSize: 10000,
+        showErrors: false,
+        wires: [
+          [],
+          [],
+          []
+        ]
+      },
+      {
+        id: '5fffb0bc.0b8a5',
+        type: 'modbus-queue-info',
+        name: 'QueueInfo',
+        topic: '',
+        unitid: '',
+        queueReadIntervalTime: 100,
+        lowLowLevel: 1,
+        lowLevel: 2,
+        highLevel: 3,
+        highHighLevel: 4,
+        server: '1e3ac4ea.86fa7b',
+        errorOnHighLevel: false,
+        showStatusActivities: true,
+        wires: [
+          ['h1']
+        ]
+      },
+      {
+        id: 'ae473c43.3e7938',
+        type: 'inject',
+        name: '',
+        topic: '',
+        payload: '',
+        payloadType: 'date',
+        repeat: 2,
+        crontab: '',
+        once: false,
+        onceDelay: 1,
+        wires: [
+          [
+            '5fffb0bc.0b8a5'
+          ]
+        ]
+      },
+      { id: 'h1', type: 'helper' },
+      {
+        id: '1e3ac4ea.86fa7b',
+        type: 'modbus-client',
+        z: '',
+        name: 'ModbsuFlexServer',
+        clienttype: 'tcp',
+        bufferCommands: true,
+        stateLogEnabled: true,
+        parallelUnitIdsAllowed: true,
+        tcpHost: '127.0.0.1',
+        tcpPort: '7502',
+        tcpType: 'DEFAULT',
+        serialPort: '/dev/ttyUSB',
+        serialType: 'RTU-BUFFERD',
+        serialBaudrate: '9600',
+        serialDatabits: '8',
+        serialStopbits: '1',
+        serialParity: 'none',
+        serialConnectionDelay: '100',
+        unit_id: '1',
+        commandDelay: '1',
+        clientTimeout: '100',
+        reconnectTimeout: 200
+      }], function () {
+        const h1 = helper.getNode('h1')
+        h1.on('input', function (msg) {
+          done()
+        })
+        const queueNode = helper.getNode('5fffb0bc.0b8a5')
+        queueNode.receive({ payload: { resetQueue: true } })
       }, function () {
         helper.log('function callback')
       })
@@ -301,10 +393,10 @@ describe('Queue Info node Testing', function () {
         topic: '',
         payload: '',
         payloadType: 'date',
-        repeat: '0.2',
+        repeat: 0.3,
         crontab: '',
         once: true,
-        onceDelay: 0.1,
+        onceDelay: 0.2,
         wires: [
           [
             '5fffb0bc.0b8a5'