+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# [5.9.0](https://github.com/biancoroyal/node-red-contrib-modbus/compare/v5.8.0...v5.9.0) (2020-04-09)
+# [5.10.0](https://github.com/biancoroyal/node-red-contrib-modbus/compare/v5.8.0...v5.10.0) (2020-04-10)
 
 
 ### Bug Fixes
@@ -6,6 +6,7 @@
 * FSM double take on running in the same states with multiple consumer ([5072e40](https://github.com/biancoroyal/node-red-contrib-modbus/commit/5072e407961ba5145ea9d50bc742c7494ba482b4))
 * HTTP messaging did not work ([8e87053](https://github.com/biancoroyal/node-red-contrib-modbus/commit/8e870534f14c0f20f7445bfe214c790b8124e7fd))
 * improve register and de-register for issue  [#165](https://github.com/biancoroyal/node-red-contrib-modbus/issues/165)  ([c68d59b](https://github.com/biancoroyal/node-red-contrib-modbus/commit/c68d59b41a8081bb9a345165098e360b787e2336))
+* issue [#187](https://github.com/biancoroyal/node-red-contrib-modbus/issues/187) - listener remove for partial deploy ([5220d27](https://github.com/biancoroyal/node-red-contrib-modbus/commit/5220d2712c357fd385feb860fe79587e1f828793))
 * read node should not use basic status handling cause of the own interval ([dea606f](https://github.com/biancoroyal/node-red-contrib-modbus/commit/dea606f8d7bcf15c385f0b9336d195e82100126a))
 * remove all listeners was lost in code ([92ada63](https://github.com/biancoroyal/node-red-contrib-modbus/commit/92ada63499a9bd0683817308cbb3f3c27393e4bb))
 * test timeout passing ([0dcf966](https://github.com/biancoroyal/node-red-contrib-modbus/commit/0dcf966d021d7abc7726976bec8be84d7d1b968b))
@@ -17,6 +18,7 @@
 * **getter:** empty msg on catch error ([c8ac4a6](https://github.com/biancoroyal/node-red-contrib-modbus/commit/c8ac4a675ce17ca1992a1018ad113652cd5b97a4))
 * **write:** empty msg on fail as the other node ([d6fd30b](https://github.com/biancoroyal/node-red-contrib-modbus/commit/d6fd30b20cf638b824852cad4a66ef42de6a07da))
 * new status update concept for nodes with less traffic ([eb2afba](https://github.com/biancoroyal/node-red-contrib-modbus/commit/eb2afbabac913eaa3d017735dc7a11eb9d02d60f))
+* queue info can handle all queues from one node issue [#175](https://github.com/biancoroyal/node-red-contrib-modbus/issues/175) ([1a094b2](https://github.com/biancoroyal/node-red-contrib-modbus/commit/1a094b27bf0846118e1db18ea4e567898a788e37))